requirements.txt
# Core dependencies for AI Content Detection System

# C2PA content credentials verification
c2pa-python>=0.4.0

# AI detection models
transformers>=4.30.0
torch>=2.0.0

# Image processing
pillow>=10.0.0
timm>=0.9.0  # For image models

# HTTP requests for soft binding API
requests>=2.31.0

# API server (optional - for running local C2PA API)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
